+++ qgemm/csrc/quant_gemm_backward.cu
@@ -21,6 +21,7 @@
 
 #include <cuda_runtime.h>
 #include <torch/extension.h>
+#include <ATen/cuda/CUDAContext.h>
 #include <stdint.h>
 
 #define CEIL_DIV(a, b) (((a) + (b) - 1) / (b))